+++ Runtime/Scripts/GlobalTools/Generic/ObjectHelpers_Geography.cs
@@ -35,216 +35,217 @@ namespace PolytopeSolutions.Toolset.GlobalTools.Generic {
             *  TODO: switch to using doubles for all coordinates.
             */
 
-        public delegate float GameTileSize(int zoomLevel);
-        public static readonly float EarthCircumference = 2 * 6378137.0f * Mathf.PI;
-        public static readonly float EarthRadius = 6372.7982f;
-        public static readonly float GeoMaxLatitude = 85.05112878f;
-        public static readonly float GeoMaxLongitude = 180f;
-        public static float GeoTileWidthMeters(int zoomLevel) => Geography.EarthCircumference / Mathf.Pow(2, (float)zoomLevel);
-        public static float GeoTileWidthDegrees(int zoomLevel) => (Geography.GeoMaxLongitude * 2) / Mathf.Pow(2, (float)zoomLevel);
+        public delegate double GameTileSize(int zoomLevel);
+        public static readonly double EarthCircumference = 2 * 6378137.0d * Math.PI;
+        public static readonly double EarthRadius = 6372.7982d;
+        public static readonly double GeoMaxLatitude = 85.05112878d;
+        public static readonly double GeoMaxLongitude = 180d;
+        public static double GeoTileWidthMeters(int zoomLevel) => Geography.EarthCircumference / Math.Pow(2, zoomLevel);
+        public static double GeoTileWidthDegrees(int zoomLevel) => (Geography.GeoMaxLongitude * 2) / Math.Pow(2, zoomLevel);
 
         // Geo coordinate converters
-        public static void Geo2GeoTile(this Vector3 geoCoordinate, int zoomLevel, ref Vector3 result) {
-            result.x = (float)(((double)geoCoordinate.x + 180.0d) / 360.0d * (1 << zoomLevel));
-            result.z = (float)((1.0d - Math.Log(Math.Tan((double)geoCoordinate.z * Math.PI / 180.0d) +
-                    1.0d / Math.Cos((double)geoCoordinate.z * Math.PI / 180.0d)) / Math.PI) / 2.0d * (1 << zoomLevel));
+        public static void Geo2GeoTile(this Vector3Double geoCoordinate, int zoomLevel, ref Vector3Double result) {
+            result.x = ((double)geoCoordinate.x + 180.0d) / 360.0d * (1 << zoomLevel);
+            result.z = (1.0d - Math.Log(Math.Tan(geoCoordinate.z * Math.PI / 180.0d) +
+                    1.0d / Math.Cos(geoCoordinate.z * Math.PI / 180.0d)) / Math.PI) / 2.0d * (1 << zoomLevel);
         }
-        public static void Geo2FlatGame(this Vector3 geoCoordinate, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
+        public static void Geo2FlatGame(this Vector3Double geoCoordinate, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
             geoCoordinate.Geo2GeoTile(zoomLevel, ref result);
             result.GeoTile2FlatGame(zoomLevel, baseGeoTile, gameTileSize, ref result);
         }
-        public static void Geo2SphericalGame(this Vector3 geoCoordinate, Vector3 center, float radius, ref Vector3 result) {
-            result = center + Quaternion.AngleAxis(-geoCoordinate.x, Vector3.up) *
-                    (Quaternion.AngleAxis(geoCoordinate.z, Vector3.right)
-                        * -Vector3.forward * radius
-                    );
+        public static void Geo2SphericalGame(this Vector3Double geoCoordinate, Vector3Double center, double radius, ref Vector3Double result) {
+            result = center + new Vector3Double(Quaternion.AngleAxis(-geoCoordinate.xf, Vector3.up) *
+                    (Quaternion.AngleAxis(geoCoordinate.zf, Vector3.right)
+                        * -Vector3.forward))
+                    * radius;
         }
-        public static void Geo2GameTile(this Vector3 geoCoordinate, int zoomLevel, Vector3 baseGeoTile, ref Vector3 result) {
+        public static void Geo2GameTile(this Vector3Double geoCoordinate, int zoomLevel, Vector3Double baseGeoTile, ref Vector3Double result) {
             geoCoordinate.Geo2GeoTile(zoomLevel, ref result);
             result.GeoTile2GameTile(baseGeoTile, ref result);
         }
-        public static void Geo2GameUV(this Vector3 geoCoordinate, int zoomLevel, ref Vector3 result) {
+        public static void Geo2GameUV(this Vector3Double geoCoordinate, int zoomLevel, ref Vector3Double result) {
             geoCoordinate.Geo2GeoTile(zoomLevel, ref result);
             result.GeoTile2GameUV(ref result);
         }
 
         // GeoTile converters
-        public static void GeoTileParent(this Vector3 geoTile, ref Vector3 result) {
-            result = geoTile / 2f;
+        public static void GeoTileParent(this Vector3Double geoTile, ref Vector3Double result) {
+            result = geoTile / 2;
         }
-        public static void GeoTileChildren(this Vector3 geoTile, ref Vector3[] result) {
+        public static void GeoTileChildren(this Vector3Double geoTile, ref Vector3Double[] result) {
             if (result == null || result.Length != 4)
-                result = new Vector3[4];
-            Vector3 geoTileFloored = geoTile.FloorToInt();
-            result[0] = geoTileFloored * 2 + Vector3Int.zero;
-            result[1] = geoTileFloored * 2 + Vector3Int.right;
-            result[2] = geoTileFloored * 2 + Vector3Int.forward;
-            result[3] = geoTileFloored * 2 + Vector3Int.right + Vector3Int.forward;
+                result = new Vector3Double[4];
+            Vector3Double geoTileFloored = geoTile.FloorToInt();
+            result[0] = geoTileFloored * 2 + Vector3Double.zero;
+            result[1] = geoTileFloored * 2 + Vector3Double.right;
+            result[2] = geoTileFloored * 2 + Vector3Double.forward;
+            result[3] = geoTileFloored * 2 + Vector3Double.right + Vector3Double.forward;
         }
-        public static void GeoTileSiblings(this Vector3 geoTile, ref Vector3[] result) {
+        public static void GeoTileSiblings(this Vector3Double geoTile, ref Vector3Double[] result) {
             if (result == null || result.Length != 3)
-                result = new Vector3[3];
-            Vector3 geoTileFloored = geoTile.FloorToInt();
-            Vector3 parent = Vector3.zero;
+                result = new Vector3Double[3];
+            Vector3Double geoTileFloored = geoTile.FloorToInt();
+            Vector3Double parent = Vector3Double.zero;
             geoTile.GeoTileParent(ref parent);
             if (parent.x.Fraction() >= 0.5f && parent.z.Fraction() >= 0.5f) {
-                result[0] = geoTileFloored - Vector3Int.right;
-                result[1] = geoTileFloored - Vector3Int.forward;
-                result[2] = geoTileFloored - Vector3Int.right - Vector3Int.forward;
+                result[0] = geoTileFloored - Vector3Double.right;
+                result[1] = geoTileFloored - Vector3Double.forward;
+                result[2] = geoTileFloored - Vector3Double.right - Vector3Double.forward;
             }
             else if (parent.x.Fraction() < 0.5f && parent.z.Fraction() >= 0.5f) {
-                result[0] = geoTileFloored + Vector3Int.right;
-                result[1] = geoTileFloored - Vector3Int.forward;
-                result[2] = geoTileFloored + Vector3Int.right - Vector3Int.forward;
+                result[0] = geoTileFloored + Vector3Double.right;
+                result[1] = geoTileFloored - Vector3Double.forward;
+                result[2] = geoTileFloored + Vector3Double.right - Vector3Double.forward;
             }
             else if (parent.x.Fraction() < 0.5f && parent.z.Fraction() < 0.5f) {
-                result[0] = geoTileFloored + Vector3Int.right;
-                result[1] = geoTileFloored + Vector3Int.forward;
-                result[2] = geoTileFloored + Vector3Int.right + Vector3Int.forward;
+                result[0] = geoTileFloored + Vector3Double.right;
+                result[1] = geoTileFloored + Vector3Double.forward;
+                result[2] = geoTileFloored + Vector3Double.right + Vector3Double.forward;
             }
             else {
-                result[0] = geoTileFloored - Vector3Int.right;
-                result[1] = geoTileFloored + Vector3Int.forward;
-                result[2] = geoTileFloored - Vector3Int.right + Vector3Int.forward;
+                result[0] = geoTileFloored - Vector3Double.right;
+                result[1] = geoTileFloored + Vector3Double.forward;
+                result[2] = geoTileFloored - Vector3Double.right + Vector3Double.forward;
             }
         }
-        public static void GeoTile2Geo(this Vector3 geoTile, int zoomLevel, ref Vector3 result) {
-            double n = Math.PI - ((2.0d * Math.PI * (double)geoTile.z) / (1 << zoomLevel));
-            result.x = (float)(((double)geoTile.x / (1 << zoomLevel) * 360.0d) - 180.0d);
-            result.z = (float)(180.0d / Math.PI * Math.Atan(Math.Sinh(n)));
+        public static void GeoTile2Geo(this Vector3Double geoTile, int zoomLevel, ref Vector3Double result) {
+            double n = Math.PI - ((2.0d * Math.PI * geoTile.z) / (1 << zoomLevel));
+            result.x = geoTile.x / (1 << zoomLevel) * 360.0d - 180.0d;
+            result.z = 180.0d / Math.PI * Math.Atan(Math.Sinh(n));
         }
-        public static void GeoTile2FlatGame(this Vector3 geoTile, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
-            float zoomGameTileSize = gameTileSize(zoomLevel);
+        public static void GeoTile2FlatGame(this Vector3Double geoTile, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
+            double zoomGameTileSize = gameTileSize(zoomLevel);
             result.x = (-baseGeoTile.x + geoTile.x) * zoomGameTileSize;
             result.z = (baseGeoTile.z - geoTile.z) * zoomGameTileSize;
         }
-        public static void GeoTile2SphericalGame(this Vector3 geoTile, int zoomLevel, Vector3 center, float radius, ref Vector3 result) {
+        public static void GeoTile2SphericalGame(this Vector3Double geoTile, int zoomLevel, Vector3Double center, double radius, ref Vector3Double result) {
             geoTile.GeoTile2Geo(zoomLevel, ref result);
             result.Geo2SphericalGame(center, radius, ref result);
         }
-        public static void GeoTile2GameTile(this Vector3 geoTile, Vector3 baseGeoTile, ref Vector3 result) {
-            result.x = -Mathf.FloorToInt(baseGeoTile.x) + geoTile.x;
-            result.z = Mathf.FloorToInt(baseGeoTile.z) - geoTile.z + 1;
+        public static void GeoTile2GameTile(this Vector3Double geoTile, Vector3Double baseGeoTile, ref Vector3Double result) {
+            result.x = -Math.Floor(baseGeoTile.x) + geoTile.x;
+            result.z = Math.Ceiling(baseGeoTile.z) - geoTile.z + 1;
         }
-        public static void GeoTile2GameUV(this Vector3 geoTile, ref Vector3 result) {
+        public static void GeoTile2GameUV(this Vector3Double geoTile, ref Vector3Double result) {
             result.x = geoTile.x.Fraction();
             result.z = 1 - geoTile.z.Fraction();
         }
-        public static void GeoTile2GeoTileIndex(this Vector3 geoTile, ref Vector3Int result) {
-            result.x = Mathf.FloorToInt(geoTile.x);
-            result.z = Mathf.FloorToInt(geoTile.z);
+        public static void GeoTile2GeoTileIndex(this Vector3Double geoTile, ref Vector3Int result) {
+            result.x = Mathf.FloorToInt(geoTile.xf);
+            result.z = Mathf.FloorToInt(geoTile.zf);
         }
-        public static void GeoTileIndex2Geo(this Vector3Int geoTileIndex, int zoomLevel, ref Vector3 result) {
-            (geoTileIndex + Vector3.right*0.5f + Vector3.forward*0.5f).GeoTile2Geo(zoomLevel, ref result);
+        public static void GeoTileIndex2Geo(this Vector3Int geoTileIndex, int zoomLevel, ref Vector3Double result) {
+            (new Vector3Double(geoTileIndex + Vector3.right*0.5f + Vector3.forward*0.5f)).GeoTile2Geo(zoomLevel, ref result);
         }
-        public static void GeoTileIndex2FlatGame(this Vector3Int geoTileIndex, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
-            (geoTileIndex + Vector3.right * 0.5f + Vector3.forward * 0.5f).GeoTile2FlatGame(zoomLevel, baseGeoTile, gameTileSize, ref result);
+        public static void GeoTileIndex2FlatGame(this Vector3Int geoTileIndex, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
+            (new Vector3Double(geoTileIndex + Vector3.right * 0.5f + Vector3.forward * 0.5f)).GeoTile2FlatGame(zoomLevel, baseGeoTile, gameTileSize, ref result);
         }
-        public static void GeoTileIndex2SphericalGame(this Vector3Int geoTileIndex, int zoomLevel, Vector3 center, float radius, ref Vector3 result) {
+        public static void GeoTileIndex2SphericalGame(this Vector3Int geoTileIndex, int zoomLevel, Vector3Double center, double radius, ref Vector3Double result) {
             geoTileIndex.GeoTileIndex2Geo(zoomLevel, ref result);
             result.Geo2SphericalGame(center, radius, ref result);
         }
-        public static void GeoTileIndex2GameTileIndex(this Vector3Int geoTileIndex, Vector3 baseGeoTile, ref Vector3Int result) {
-            result.x = Mathf.FloorToInt(-Mathf.FloorToInt(baseGeoTile.x) + geoTileIndex.x + 0.5f);
-            result.z = Mathf.FloorToInt(Mathf.FloorToInt(baseGeoTile.z) - geoTileIndex.z + 0.5f);
+        public static void GeoTileIndex2GameTileIndex(this Vector3Int geoTileIndex, Vector3Double baseGeoTile, ref Vector3Int result) {
+            result.x = Mathf.FloorToInt(-Mathf.FloorToInt(baseGeoTile.xf) + geoTileIndex.x + 0.5f);
+            result.z = Mathf.FloorToInt(Mathf.FloorToInt(baseGeoTile.zf) - geoTileIndex.z + 0.5f);
         }
         
         // Flat game coordinate converters
-        public static void FlatGame2GameTile(this Vector3 flatGameCoordinate, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
+        public static void FlatGame2GameTile(this Vector3Double flatGameCoordinate, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
             // Calculate game tile index accounting for the in-game offset of base tile.
-            float zoomGameTileSize = gameTileSize(zoomLevel);
+            double zoomGameTileSize = gameTileSize(zoomLevel);
             result.x = baseGeoTile.x.Fraction() + flatGameCoordinate.x / zoomGameTileSize;
             result.z = -baseGeoTile.z.Fraction() + 1f + flatGameCoordinate.z / zoomGameTileSize;
         }
-        public static void FlatGame2GeoTile(this Vector3 flatGameCoordinate, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
+        public static void FlatGame2GeoTile(this Vector3Double flatGameCoordinate, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
             // Round down game tile to index relative to base offset.
-            float zoomGameTileSize = gameTileSize(zoomLevel);
+            double zoomGameTileSize = gameTileSize(zoomLevel);
             result.x = baseGeoTile.x + flatGameCoordinate.x / zoomGameTileSize;
             result.z = baseGeoTile.z - flatGameCoordinate.z / zoomGameTileSize;
         }
-        public static void FlatGame2Geo(this Vector3 flatGameCoordinate, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
+        public static void FlatGame2Geo(this Vector3Double flatGameCoordinate, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
             flatGameCoordinate.FlatGame2GeoTile(zoomLevel, baseGeoTile, gameTileSize, ref result);
             result.GeoTile2Geo(zoomLevel, ref result);
         }
-        public static void FlatGame2GameUV(this Vector3 flatGameCoordinate, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
+        public static void FlatGame2GameUV(this Vector3Double flatGameCoordinate, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
             flatGameCoordinate.FlatGame2GameTile(zoomLevel, baseGeoTile, gameTileSize, ref result);
             result.GameTile2GameUV(ref result);
         }
         
         // Spherical game coordinate converters
-        public static void SphericalGame2GameTile(this Vector3 sphericalGameCoordinate, Vector3 center, int zoomLevel, Vector3 baseGeoTile, ref Vector3 result) {
+        public static void SphericalGame2GameTile(this Vector3Double sphericalGameCoordinate, Vector3Double center, int zoomLevel, Vector3Double baseGeoTile, ref Vector3Double result) {
             sphericalGameCoordinate.SphericalGame2Geo(center, ref result);
             result.Geo2GameTile(zoomLevel, baseGeoTile, ref result);
         }
-        public static void SphericalGame2GeoTile(this Vector3 sphericalGameCoordinate, Vector3 center, int zoomLevel, ref Vector3 result) {
+        public static void SphericalGame2GeoTile(this Vector3Double sphericalGameCoordinate, Vector3Double center, int zoomLevel, ref Vector3Double result) {
             sphericalGameCoordinate.SphericalGame2Geo(center, ref result);
             result.Geo2GeoTile(zoomLevel, ref result);
         }
-        public static void SphericalGame2Geo(this Vector3 sphericalGameCoordinate, Vector3 center, ref Vector3 result) {
+        public static void SphericalGame2Geo(this Vector3Double sphericalGameCoordinate, Vector3Double center, ref Vector3Double result) {
             // Check if axis is vertical - then just give (0, 90) or (0, -90).
-            float verticalAllignment = Vector3.Dot((sphericalGameCoordinate - center).normalized, Vector3.up);
+            Vector3 offsetNormalized = (sphericalGameCoordinate - center).normalized.ToVector3();
+            float verticalAllignment = Vector3.Dot(offsetNormalized, Vector3.up);
             if (Mathf.Abs(Mathf.Abs(verticalAllignment) - 1f) < float.Epsilon) { 
-                result = Mathf.Sign(verticalAllignment) * Vector3.up * 90f;
+                result = Vector3Double.up * Mathf.Sign(verticalAllignment) * 90f;
                 return;
             }
             // Compute normal to plane with vertical axis.
-            Vector3 verticalPlaneNormal = Vector3.Cross(Vector3.up, (sphericalGameCoordinate - center));
+            Vector3 verticalPlaneNormal = Vector3.Cross(Vector3.up, offsetNormalized);
             // Compute horizontal projection of coordinate.
             Vector3 horizontalAxis = Vector3.Cross(verticalPlaneNormal, Vector3.up);
             // Compute angles within the vertical plane for y and within projected horizontal plane for x
-            float anglePitch = -Vector3.SignedAngle(horizontalAxis, (sphericalGameCoordinate - center), verticalPlaneNormal);
+            float anglePitch = -Vector3.SignedAngle(horizontalAxis, offsetNormalized, verticalPlaneNormal);
             float angleYaw = -Vector3.SignedAngle(-Vector3.forward, horizontalAxis, Vector3.up);
 
             result.x = angleYaw;
             result.z = anglePitch;
         }
-        public static void SphericalGame2GameUV(this Vector3 sphericalGameCoordinate, Vector3 center, int zoomLevel, ref Vector3 result) {
+        public static void SphericalGame2GameUV(this Vector3Double sphericalGameCoordinate, Vector3Double center, int zoomLevel, ref Vector3Double result) {
             sphericalGameCoordinate.SphericalGame2Geo(center, ref result);
             result.Geo2GameUV(zoomLevel, ref result);
         }
 
         // GameTile converters
-        public static void GameTile2Geo(this Vector3 gameTile, int zoomLevel, Vector3 baseGeoTile, ref Vector3 result) {
+        public static void GameTile2Geo(this Vector3Double gameTile, int zoomLevel, Vector3Double baseGeoTile, ref Vector3Double result) {
             gameTile.GameTile2GeoTile(baseGeoTile, ref result);
             result.GeoTile2Geo(zoomLevel, ref result);
         }
-        public static void GameTile2GeoTile(this Vector3 gameTile, Vector3 baseGeoTile, ref Vector3 result) {
-            result.x = Mathf.FloorToInt(baseGeoTile.x) + gameTile.x; 
-            result.z = Mathf.FloorToInt(baseGeoTile.z) - gameTile.z + 1;
+        public static void GameTile2GeoTile(this Vector3Double gameTile, Vector3Double baseGeoTile, ref Vector3Double result) {
+            result.x = Math.Floor(baseGeoTile.x) + gameTile.x; 
+            result.z = Math.Floor(baseGeoTile.z) - gameTile.z + 1;
         }
-        public static void GameTile2FlatGame(this Vector3 gameTile, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
-            float zoomGameTileSize = gameTileSize(zoomLevel);
+        public static void GameTile2FlatGame(this Vector3Double gameTile, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
+            double zoomGameTileSize = gameTileSize(zoomLevel);
             result.x = (-baseGeoTile.x.Fraction() + gameTile.x) * zoomGameTileSize;
             result.z = (baseGeoTile.z.Fraction() - 1 + gameTile.z) * zoomGameTileSize;
         }
-        public static void GameTile2SphericalGame(this Vector3 gameTile, Vector3 center, int zoomLevel, Vector3 baseGeoTile, float radius, ref Vector3 result) {
+        public static void GameTile2SphericalGame(this Vector3Double gameTile, Vector3Double center, int zoomLevel, Vector3Double baseGeoTile, double radius, ref Vector3Double result) {
             gameTile.GameTile2Geo(zoomLevel, baseGeoTile, ref result);
             result.Geo2SphericalGame(center, radius, ref result);
         }
-        public static void GameTile2GameUV(this Vector3 gameTile, ref Vector3 result) {
+        public static void GameTile2GameUV(this Vector3Double gameTile, ref Vector3Double result) {
             result.x = gameTile.x.Fraction();
             result.z = gameTile.z.Fraction();
             // Handle negative tile indices
             result.x = (result.x > 0) ? result.x : 1 + result.x;
             result.z = (result.z > 0) ? result.z : 1 + result.z;
         }
-        public static void GameTile2GameTileIndex(this Vector3 gameTile, ref Vector3Int result) {
-            result.x = Mathf.FloorToInt(gameTile.x);
-            result.z = Mathf.FloorToInt(gameTile.z);
+        public static void GameTile2GameTileIndex(this Vector3Double gameTile, ref Vector3Int result) {
+            result.x = Mathf.FloorToInt(gameTile.xf);
+            result.z = Mathf.FloorToInt(gameTile.zf);
         }
-        public static void GameTileIndex2Geo(this Vector3Int gameTileIndex, int zoomLevel, Vector3 baseGeoTile, ref Vector3 result) {
-            (gameTileIndex + Vector3.right * 0.5f + Vector3.forward * 0.5f)
+        public static void GameTileIndex2Geo(this Vector3Int gameTileIndex, int zoomLevel, Vector3Double baseGeoTile, ref Vector3Double result) {
+            (new Vector3Double(gameTileIndex + Vector3.right * 0.5f + Vector3.forward * 0.5f))
                 .GameTile2Geo(zoomLevel, baseGeoTile, ref result);
         }
-        public static void GameTileIndex2GeoTileIndex(this Vector3Int gameTileIndex, Vector3 baseGeoTile, ref Vector3Int result) {
-            result.x = Mathf.FloorToInt(Mathf.FloorToInt(baseGeoTile.x) + gameTileIndex.x + 0.5f);
-            result.z = Mathf.FloorToInt(Mathf.FloorToInt(baseGeoTile.z) - gameTileIndex.z + 0.5f);
+        public static void GameTileIndex2GeoTileIndex(this Vector3Int gameTileIndex, Vector3Double baseGeoTile, ref Vector3Int result) {
+            result.x = Mathf.FloorToInt(Mathf.FloorToInt(baseGeoTile.xf) + gameTileIndex.x + 0.5f);
+            result.z = Mathf.FloorToInt(Mathf.FloorToInt(baseGeoTile.zf) - gameTileIndex.z + 0.5f);
         }
-        public static void GameTileIndex2FlatGame(this Vector3Int gameTileIndex, int zoomLevel, Vector3 baseGeoTile, GameTileSize gameTileSize, ref Vector3 result) {
-            (gameTileIndex + Vector3.right * 0.5f + Vector3.forward * 0.5f)
+        public static void GameTileIndex2FlatGame(this Vector3Int gameTileIndex, int zoomLevel, Vector3Double baseGeoTile, GameTileSize gameTileSize, ref Vector3Double result) {
+            (new Vector3Double(gameTileIndex + Vector3.right * 0.5f + Vector3.forward * 0.5f))
                 .GameTile2FlatGame(zoomLevel, baseGeoTile, gameTileSize, ref result);
         }
-        public static void GameTileIndex2SphericalGame(this Vector3Int gameTileIndex, int zoomLevel, Vector3 baseGeoTile, Vector3 center, float radius, ref Vector3 result) {
+        public static void GameTileIndex2SphericalGame(this Vector3Int gameTileIndex, int zoomLevel, Vector3Double baseGeoTile, Vector3Double center, float radius, ref Vector3Double result) {
             gameTileIndex.GameTileIndex2Geo(zoomLevel, baseGeoTile, ref result);
             result.Geo2SphericalGame(center, radius, ref result);
         }